read_subscriptions/requirements.txt
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-subscription==3.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
